+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Upcoming bug-fix release, expected around 15 September 2026.
 
  - [#865] `CompressedImageTiler` to properly read headers in rather than try to adapt to specific algorithms during decompression (see `initCompressionOption()` method), and remove reverse traversal from row read in favour of `ZNAXIS` reverse read. (by @at88mph) 
 
+ - [#874] Fixed processing of unquantized compressed HDUs when `ZQUANTIZ` keyword is present. Quantization is applied only if `ZZERO` and `SCALE` columns are present. When they are absent `ZQUANTIZ` has no implication. (by @attipaci, thanks to @timj and @robyww)
+
 ### Changed
 
  - The latest build and runtime Maven dependencies. (by @attipaci)